+++ src/features/settings/components/networks-configs-table.tsx
@@ -5,14 +5,29 @@ import { DataTable } from '@/features/common/components/data-table'
 import { Button } from '@/features/common/components/button'
 import { Dialog, DialogContent, DialogHeader, MediumSizeDialogBody } from '@/features/common/components/dialog'
 import { useCallback, useState } from 'react'
-import { NetworkForm } from '@/features/settings/components/network-form'
+import { EditNetworkConfigForm } from '@/features/settings/components/edit-network-config-form'
+import { CreateNetworkConfigForm } from '@/features/settings/components/create-network-config-form'
 
-export function NetworksTable() {
+// TODO: delete, on small screens
+export function NetworksConfigsTable() {
+  const [createNetworkConfigDialogOpen, setCreateNetworkConfigDialogOpen] = useState(false)
   const networksConfigs = useNetworksConfigs()
   return (
     <>
       <h2>Networks</h2>
-      <DataTable columns={tableColumns} data={networksConfigs} />
+      <DataTable columns={tableColumns} data={networksConfigs} onCreateButtonClick={() => setCreateNetworkConfigDialogOpen(true)} />
+      <Dialog open={createNetworkConfigDialogOpen} onOpenChange={setCreateNetworkConfigDialogOpen} modal={true}>
+        {createNetworkConfigDialogOpen && (
+          <DialogContent className="bg-card">
+            <DialogHeader className="flex-row items-center space-y-0">
+              <h2 className="pb-0">Create Network</h2>
+            </DialogHeader>
+            <MediumSizeDialogBody>
+              <CreateNetworkConfigForm onSuccess={() => setCreateNetworkConfigDialogOpen(false)} />
+            </MediumSizeDialogBody>
+          </DialogContent>
+        )}
+      </Dialog>
     </>
   )
 }
@@ -76,7 +91,7 @@ function EditNetworkButton({ network }: { network: NetworkConfig }) {
               <h2 className="pb-0">Edit {network.name}</h2>
             </DialogHeader>
             <MediumSizeDialogBody>
-              <NetworkForm network={network} onSuccess={() => setDialogOpen(false)} />
+              <EditNetworkConfigForm network={network} onSuccess={() => setDialogOpen(false)} />
             </MediumSizeDialogBody>
           </DialogContent>
         )}
