+++ src/features/settings/form-schemas/edit-network-config-form-schema.ts
@@ -1,0 +1,10 @@
+import { zfd } from 'zod-form-data'
+import { algoServiceSchema, kmdServiceSchema } from '@/features/settings/form-schemas/algo-service-schema'
+
+export const editNetworkConfigFormSchema = zfd.formData({
+  networkId: zfd.text(),
+  name: zfd.text(),
+  indexer: algoServiceSchema,
+  algod: algoServiceSchema,
+  kmd: kmdServiceSchema,
+})
